Created TaskScreen
This consists of many changes to the older code base, much of which has been worked over the time since the last commit. The exact changes and order has been forgotten because I hadn't stayed on top of my commit history. And this is why we can't have nice things. Moving forward I plan to do better.